RavenDB-20328 fix sharded subscription drop
