+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/Subscriptions/ShardedSubscriptionsHandlerProcessorForDropSubscriptionConnection.cs
@@ -4,6 +4,7 @@ using JetBrains.Annotations;
 using Raven.Client.Exceptions.Documents.Subscriptions;
 using Raven.Server.Documents.Handlers.Processors.Subscriptions;
 using Raven.Server.Documents.Sharding.Operations;
+using Raven.Server.Documents.Subscriptions;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Utils;
 
@@ -17,34 +18,37 @@ namespace Raven.Server.Documents.Sharding.Handlers.Processors.Subscriptions
 
         protected override async ValueTask DropSubscriptionAsync(long? subscriptionId, string subscriptionName, string workerId)
         {
-            try
-            {
-                var op = new ShardedDropSubscriptionConnectionOperation(HttpContext, subscriptionName, subscriptionId, workerId);
-                await RequestHandler.ShardExecutor.ExecuteParallelForAllAsync(op);
-            }
-            finally
-            {
-                if (string.IsNullOrEmpty(subscriptionName))
-                {
-                    if (subscriptionId.HasValue == false)
-                        throw new ArgumentException("Subscription drop operation must get either subscription name or subscription task id");
-                    
-                    //get subscription name from state
-                    using (ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-                    {
-                        subscriptionName = ServerStore.Cluster.Subscriptions.GetSubscriptionNameById(context, RequestHandler.DatabaseName, subscriptionId.Value);
-                    }
+            if (subscriptionId.HasValue == false && string.IsNullOrEmpty(subscriptionName))
+                throw new ArgumentException("Subscription drop operation must get either subscription name or subscription task id");
 
-                    if (subscriptionName == null)
-                        throw new SubscriptionDoesNotExistException($"Could not find a subscription with a subscription id of {subscriptionId.Value}");
-                }
-
-                DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Stav, DevelopmentHelper.Severity.Normal, "RavenDB-19079 Make this identical to the normal EP");
-                if (RequestHandler.DatabaseContext.SubscriptionsStorage.Subscriptions.TryRemove(subscriptionId.Value, out var state))
+            if (subscriptionId.HasValue == false)
+            {
+                using (ContextPool.AllocateOperationContext(out TransactionOperationContext context))
+                using (context.OpenReadTransaction())
                 {
-                    state.Dispose();
+                    subscriptionId = ServerStore.Cluster.Subscriptions.ReadSubscriptionStateByName(context, RequestHandler.DatabaseName, subscriptionName)
+                        .SubscriptionId;
                 }
             }
+
+            if (string.IsNullOrEmpty(subscriptionName))
+            {
+                using (ContextPool.AllocateOperationContext(out TransactionOperationContext context))
+                using (context.OpenReadTransaction())
+                {
+                    subscriptionName = ServerStore.Cluster.Subscriptions.ReadSubscriptionStateById(context, RequestHandler.DatabaseName, subscriptionId.Value).SubscriptionName;
+                }
+            }
+
+            var op = new ShardedDropSubscriptionConnectionOperation(HttpContext, subscriptionName);
+            await RequestHandler.ShardExecutor.ExecuteParallelForAllAsync(op);
+            
+            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Stav, DevelopmentHelper.Severity.Normal, "RavenDB-19079 Make this identical to the normal EP");
+            if (RequestHandler.DatabaseContext.SubscriptionsStorage.Subscriptions.TryRemove(subscriptionId.Value, out var state))
+            {
+                state.Dispose();
+            }
+
             DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Stav, DevelopmentHelper.Severity.Normal, "RavenDB-19079 Handle status code");
         }
     }
